AIS-7: Implement Extraction of JWT Token

+++ src/router/v1/endpoint/chatbot.py
@@ -1,6 +1,6 @@
 import os
 
-from fastapi import APIRouter
+from fastapi import APIRouter, Header, HTTPException
 from ctransformers import AutoModelForCausalLM,AutoTokenizer
 import json
 import shutil
@@ -16,6 +16,8 @@ from core.controller.orchestration_layer.embedding_pipeline import EmbeddingPipe
 from langchain.embeddings import HuggingFaceEmbeddings
 from langchain.llms import CTransformers
 
+from src.core.controller.authentication_layer.jwt import decodeJWT
+
 llm = CTransformers(
     model=Param.LLM_MODEL_PATH,
     model_type=Param.LLM_MODEL_TYPE,
@@ -29,41 +31,56 @@ def health_check():
     return {"status":'healthy'}
 
 @router.post("/create_embedding")
-def create_embedding(file: UploadFile):
-
-    user_folder=Param.EMBEDDING_MODEL_PATH+"zhuofan.chen/"
-    if os.path.exists(user_folder):
-        shutil.rmtree(user_folder)
+def create_embedding(file: UploadFile,authorization: str = Header(None)):
+    auth=decodeJWT(authorization)
+    if (auth['valid']):
+        user_folder=Param.EMBEDDING_MODEL_PATH+auth['data']['username']+'/'
+        if os.path.exists(user_folder):
+            shutil.rmtree(user_folder)
+        else:
+            os.makedirs(user_folder)
+        file_location = f"{Param.TEMP_SAVE_PATH}/{file.filename}"
+        with open(file_location, "wb+") as file_object:
+            file_object.write(file.file.read())
+        embedding=EmbeddingPipeline(file_location,+auth['data']['username'])
+        embedding.save_db_local()
+        if os.path.isfile(file_location):
+            os.remove(file_location)
+        return APIResponse(status='success',message='Embedding Created Success')
     else:
-        os.makedirs(user_folder)
-    file_location = f"{Param.TEMP_SAVE_PATH}/{file.filename}"
-    with open(file_location, "wb+") as file_object:
-        file_object.write(file.file.read())
-    embedding=EmbeddingPipeline(file_location,'zhuofan.chen')
-    embedding.save_db_local()
-    if os.path.isfile(file_location):
-        os.remove(file_location)
-    return APIResponse(status='success',message='Embedding Created Success')
+        HTTPException(401, detail="Unauthorised")
+
+
 
 
 @router.post("/predict")
-def predict(data: PredictionRequest):
-    retriever=load_embedding(Param.EMBEDDING_SAVE_PATH+'zhuofan.chen'+"/")
-    chain = ConversationalRetrievalChain.from_llm(llm=llm, retriever=retriever.as_retriever())
-    result=LLM(chain, llm, retriever).predict(data.query,data.chat_history)
+def predict(data: PredictionRequest,authorization: str = Header(None)):
+    auth=decodeJWT(authorization)
+    if (auth['valid']):
+        retriever=load_embedding(Param.EMBEDDING_SAVE_PATH+auth['data']['username']+'/')
+        chain = ConversationalRetrievalChain.from_llm(llm=llm, retriever=retriever.as_retriever())
+        result=LLM(chain, llm, retriever).predict(data.query,data.chat_history)
 
-    return APIResponse(status='success',message=result)
+        return APIResponse(status='success',message=result)
+    else:
+        HTTPException(401, detail="Unauthorised")
+
 
 
 @router.post("/feedback")
-def feedback(data: FeedbackRequest):
-    with open(Param.FEEDBACK_LOG_FILE, 'a') as log_file:
-        log_file.write(f"User_Timestamp: {data.user_timestamp} | User_Input: {data.chat_history[len(data.chat_history)-2]}\n")
-        log_file.write(f"Bot_Timestamp: {data.bot_timestamp} | Bot_Response: {data.chat_history[-1]}\n")
-        log_file.write(f"Feedback_Timestamp: {data.feedback_timestamp} | User_Feedback: {data.feedback}\n")
-        log_file.write("=" * 50 + "\n")
+def feedback(data: FeedbackRequest,authorization: str = Header(None)):
+    auth=decodeJWT(authorization)
+    if (auth['valid']):
+        with open(Param.FEEDBACK_LOG_FILE, 'a') as log_file:
+            log_file.write(f"User_Timestamp: {data.user_timestamp} | User_Input: {data.chat_history[len(data.chat_history)-2]}\n")
+            log_file.write(f"Bot_Timestamp: {data.bot_timestamp} | Bot_Response: {data.chat_history[-1]}\n")
+            log_file.write(f"Feedback_Timestamp: {data.feedback_timestamp} | User_Feedback: {data.feedback}\n")
+            log_file.write("=" * 50 + "\n")
 
-    return APIResponse(status='success',message='Feedback Noted')
+        return APIResponse(status='success',message='Feedback Noted')
+    else:
+        HTTPException(401, detail="Unauthorised")
+
